Added XcfaMetadata

+++ subprojects/xcfa/xcfa-cli/src/test/resources/llvm/example_addition.c
@@ -1,6 +1,6 @@
 int main() {
-    volatile float a = 2;
-    volatile double b = 3;
+    volatile int a = 2;
+    volatile int b = 3;
     int c = a+b;
     return c;
 }
